stage
added game logs for every team. completed run_model to just include a url from baseball reference to run every game for the day. Also changed the model to include inputs for if the pitcher needs an input for estimated IP, and can enter the pitcher name if name not same as reference sheet.